requirements.txt
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
statsmodels==0.13.5
streamlit==1.17.0
